damageVerifierMultiAgent/requirements.txt
uagents
python-dotenv==1.0.0
